Layout and paint move to logical pixels, glyphs rasterize at device
resolution, and the translation-layer math is pinned at integer and
fractional scales. Verified on-device at 200% / 150% / 125% plus a live
scale-factor change, with no widget, caret, or text-crispness regressions.

+++ crates/shroud_widgets/src/dropdown.rs
@@ -371,8 +371,8 @@ impl Widget for Dropdown {
             for glyph in &shaped.glyphs {
                 if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                     ctx.draw_glyph(
-                        text_x as i32 + glyph.x,
-                        text_y as i32 + glyph.y,
+                        text_x + glyph.x,
+                        text_y + glyph.y,
                         image,
                         text_color,
                         glyph.cache_key,
@@ -387,8 +387,8 @@ impl Widget for Dropdown {
         for glyph in &chevron_shaped.glyphs {
             if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                 ctx.draw_glyph(
-                    chev_x as i32 + glyph.x,
-                    chev_y as i32 + glyph.y,
+                    chev_x + glyph.x,
+                    chev_y + glyph.y,
                     image,
                     text_color,
                     glyph.cache_key,
