+++ crates/shroud_widgets/src/input.rs
@@ -2227,8 +2227,8 @@ impl Widget for Input {
                     for glyph in &shaped.glyphs {
                         if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                             ctx.draw_glyph(
-                                text_x as i32 + glyph.x,
-                                text_y as i32 + glyph.y,
+                                text_x + glyph.x,
+                                text_y + glyph.y,
                                 image,
                                 placeholder_color,
                                 glyph.cache_key,
@@ -2275,8 +2275,8 @@ impl Widget for Input {
                 for glyph in &shaped.glyphs {
                     if let Some(image) = ctx.text_engine.rasterize(glyph.cache_key) {
                         ctx.draw_glyph(
-                            text_x as i32 + glyph.x,
-                            text_y as i32 + glyph.y,
+                            text_x + glyph.x,
+                            text_y + glyph.y,
                             image,
                             glyph.color.unwrap_or(text_color),
                             glyph.cache_key,
